+++ spec/shared_context.rb
@@ -39,6 +39,14 @@ RSpec.shared_context 'shared context Image Streamer', a: :b do
   end
 end
 
+RSpec.shared_context 'shared context Hypervisor Manager', a: :b do
+  before :each do
+    api_version = 800
+    options = { url: 'https://oneview.example.com', user: 'Administrator', password: 'secret123', api_version: api_version }
+    @client = OneviewSDK::Client.new(options)
+  end
+end
+
 RSpec.shared_context 'integration context', a: :b do
   # Context for integration testing:
   # WARNING: Communicates with & modifies a real instance.